move load reducer into App folder and update files
