+++ src/store/rootReducer.js
@@ -1,6 +1,6 @@
 import { combineReducers } from 'redux';
 import { categoryReducer } from '../category/reducer';
-import { loadResponse, loadSpinner, loadError } from '../load/reducer';
+import { loadResponse, loadSpinner, loadError } from '../App/reducer';
 
 const rootReducer = combineReducers({
   loadResponse,
